priceinshowpage

+++ db/migrate/20240326131333_create_bookings.rb
@@ -4,8 +4,8 @@ class CreateBookings < ActiveRecord::Migration[7.0]
       t.references :user, foreign_key: true
       t.references :flat, foreign_key: true
 
-      t.decimal :price_per_night, precision: 10, scale: 2
-      t.decimal :price_in_total, precision: 10, scale: 2
+      t.float :price_per_night, precision: 10, scale: 2
+      t.float :price_in_total, precision: 10, scale: 2
       t.date :checkin_date
       t.date :checkout_date
       t.integer :guests
